Move classes to new folder and update usings
Move classes from the differential geometry folder to the general relativity folder if they pertained to general relativity
