+++ GeneralRelativity/Signature.cs
@@ -1,0 +1,8 @@
+﻿using Physics.NET.GeneralRelativity;
+
+namespace Physics.NET.Mathematics.DifferentialGeometry
+{
+    public class Signature<T> : ISignature where T : class, ISignature
+    {
+    }
+}
