+++ GeneralRelativity/Spacelike.cs
@@ -1,0 +1,8 @@
+﻿using Physics.NET.GeneralRelativity;
+
+namespace Physics.NET.Mathematics.DifferentialGeometry
+{
+    public class Spacelike : ISignature
+    {
+    }
+}
