+++ GeneralRelativity/Timelike.cs
@@ -1,0 +1,6 @@
+﻿namespace Physics.NET.GeneralRelativity
+{
+    public class Timelike : ISignature
+    {
+    }
+}
