--- Mathematics/DifferentialGeometry/Signature.cs
@@ -1,6 +1,0 @@
-﻿namespace Physics.NET.Mathematics.DifferentialGeometry
-{
-    public class Signature<T> : ISignature where T : class, ISignature
-    {
-    }
-}
