--- Mathematics/DifferentialGeometry/Timelike.cs
@@ -1,6 +1,0 @@
-﻿namespace Physics.NET.Mathematics.DifferentialGeometry
-{
-    public class Timelike : ISignature
-    {
-    }
-}
